requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
boto3==1.9.173
botocore==1.12.173
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
confusable-homoglyphs==3.2.0
cycler==0.10.0
defusedxml==0.6.0
Django==2.1.1
django-crispy-forms==1.7.2
django-registration==3.0.1
django-storages==1.7.1
docutils==0.14
ffmpeg==1.4
ffmpy==0.2.2
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
kiwisolver==1.1.0
kombu==4.6.5
matplotlib==3.1.1
more-itertools==7.2.0
mutagen==1.42.0
mysqlclient==1.4.2.post1
numpy==1.16.4
oauthlib==3.0.1
opencv-python==4.1.0.25
Pillow==6.1.0
psycopg2==2.8.3
pycurl==7.43.0.1
pydub==0.23.1
PyJWT==1.7.1
pyparsing==2.4.1.1
PyQt5==5.13.0
PyQt5-sip==4.19.18
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
selenium==3.141.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stripe==2.30.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
